+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Main changes:
 
 Complete list:
 - Fix area selection, `default_coords` are deduced from the dataset (#2768)
+- Expose fldstat methods through `AquaAccessor` (#2770)
 - Allow level selection for non-fdb sources (#2731)
 - `frequency_string_to_pandas` function can now handle frequencies with a numerical prefix, such as "3hourly" (#2730)